requirements.txt
drf-spectacular
django
djangorestframework
django-filter
djangorestframework-simplejwt
